pyproject.toml
[tool.poetry]

name = "django-admin-action-tools"
authors = [
    "Thu Trang Pham <thuutrangpham@gmail.com>",
    "jeanloup.monnier <jean-loup.monnier@spikeelabs.fr>",
]
maintainers = ["jeanloup.monnier <jean-loup.monnier@spikeelabs.fr>"]
version = "0.1.0"
description = "Tools for django admin"
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/SpikeeLabs/django-admin-action-tools"
classifiers = [
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3 :: Only",
]
packages = [{ include = "admin_action_tools" }]

[tool.poetry.dependencies]
python = "^3.8"
Django = ">=3.2, <= 4.2"

[tool.poetry.dev-dependencies]
factory-boy = "^3.0"
coverage = "^5.4"
pytest = "^6.2"
pytest-django = "^4.1"
readme-renderer = "^28.0"
twine = "^3.3"
coveralls = "^3.0"
selenium = "^4.0"
django-storages = "^1.11"
boto3 = "^1.17"
wheel = "^0.37"
localstack = "^1.1.0"
pylint = "^2.12"
pylint-django = "^2.5"
black = "22.3.0"
pre-commit = "^2.16"
pylama = "^8.4.1"
Pillow = "^9.2.0"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
(
    /(
        \.git
        |\.tox
        |migrations
    )/
)
'''
include = '\.pyi?$'
line-length = 119

[tool.pylama]
format = "pycodestyle"
linters = "pycodestyle,pyflakes,pylint"
max_line_length = 119
skip = ".pytest_cache,.venv/*,*/migrations/*,tests/*,*/tests/*,docs/*"

[tool.pylama.linter.pycodestyle]
max_line_length = 119

[tool.pylama.linter.pylint]
disable = "W0108,W0511,W0602,W0603,W0703,C0206,C0209,C0114,C0115,C0116,R0903,R0913,R0914,R0901,E1101,E1130,E1136"
# Ignored rules:
#   - W0108: Lambda may not be necessary
#   - W0511: fixme, todo
#   - W0602: global-variable-not-assigned
#   - W0603: global-statement
#   - W0703: Catching too general exception
#   - C0114: missing-module-docstring
#   - C0115: missing-class-docstring
#   - C0116: missing-function-docstring
#   - C0206: consider-using-dict-items
#   - C0209: consider-using-f-string
#   - R0903: too-few-public-methods
#   - R0913: too-many-arguments
#   - R0914: too-many-locals
#   - R0901: max-parents: Maximum number of parents for a class
#   - E1101: generated-members
#            List of members which are set dynamically and missed by pylint inference system,
#            and so shouldn't trigger E1101 when accessed.
#   - E1130: invalid-unary-operand-type
#            Emitted when a unary operand is used on an object which does not support this type of operation.
#   - E1136: unsubscriptable-object
#            Value '%s' is unsubscriptable Emitted when a subscripted value doesn't support subscription


[tool.coverage.run]
relative_files = true
omit = ["admin_action_tools/tests/*"]
source = ["admin_action_tools"]
branch = true

[tool.coverage.report]
fail_under = 100

[tool.pytest.ini_options]
testpaths = "admin_action_tools/tests"
DJANGO_SETTINGS_MODULE = "tests.test_project.settings.test"
addopts = "--doctest-modules -ra -l --tb=short --show-capture=all --color=yes"
